Find all symbols used in a function or language object.

Find all symbols used in a function or language object.

find_symbols(x)

Arguments

x
A language object or function.

Value

A character vector of the symbols used in the object.

Examples

# Find all symbols used in the lm function
find_symbols(utils::install.packages)

#>   [1] "::"                                 ":::"                               
#>   [3] "!"                                  "!="                                
#>   [5] "..."                                ".find_install_order"               
#>   [7] ".install.macbinary"                 ".install.winbinary"                
#>   [9] ".libPaths"                          ".Library"                          
#>  [11] ".make_dependency_list"              ".Platform"                         
#>  [13] ".R_LIBS"                            ".standard_regexps"                 
#>  [15] ".TkUp"                              "("                                 
#>  [17] "["                                  "[["                                
#>  [19] "{"                                  "&"                                 
#>  [21] "&&"                                 "%in%"                              
#>  [23] "<"                                  "<-"                                
#>  [25] "=="                                 ">"                                 
#>  [27] "|"                                  "||"                                
#>  [29] "$"                                  "action"                            
#>  [31] "aDL"                                "all"                               
#>  [33] "ans"                                "any"                               
#>  [35] "args"                               "args0"                             
#>  [37] "as.numeric_version"                 "askYesNo"                          
#>  [39] "av"                                 "av1"                               
#>  [41] "av2"                                "available"                         
#>  [43] "available_packages_filters_db"      "available_packages_filters_default"
#>  [45] "available.packages"                 "basename"                          
#>  [47] "bins"                               "binvers"                           
#>  [49] "c"                                  "capabilities"                      
#>  [51] "cat"                                "cbind"                             
#>  [53] "character"                          "clean"                             
#>  [55] "close"                              "cmd"                               
#>  [57] "cmd0"                               "colnames"                          
#>  [59] "config"                             "configure.args"                    
#>  [61] "configure.vars"                     "conn"                              
#>  [63] "contrib.url"                        "contriburl"                        
#>  [65] "contriburl2"                        "cwd"                               
#>  [67] "data.frame"                         "dependencies"                      
#>  [69] "deps"                               "destdir"                           
#>  [71] "destfile"                           "df"                                
#>  [73] "dir"                                "dir.create"                        
#>  [75] "dir.exists"                         "dirname"                           
#>  [77] "dirs"                               "DL"                                
#>  [79] "download.file"                      "download.packages"                 
#>  [81] "e"                                  "endsWith"                          
#>  [83] "env"                                "f"                                 
#>  [85] "f2"                                 "failed"                            
#>  [87] "fil"                                "file"                              
#>  [89] "file.access"                        "file.copy"                         
#>  [91] "file.exists"                        "file.path"                         
#>  [93] "file.remove"                        "files"                             
#>  [95] "filters"                            "flush.console"                     
#>  [97] "fn"                                 "for"                               
#>  [99] "found"                              "foundpkgs"                         
#> [101] "function"                           "get_install_opts"                  
#> [103] "get_package_name"                   "get_timeout"                       
#> [105] "getConfigureArgs"                   "getConfigureVars"                  
#> [107] "getDependencies"                    "getDeps"                           
#> [109] "getOption"                          "gettext"                           
#> [111] "gettextf"                           "getwd"                             
#> [113] "grep"                               "grepl"                             
#> [115] "gsub"                               "GUI"                               
#> [117] "hasArchs"                           "hasSrc"                            
#> [119] "have_spaces"                        "i"                                 
#> [121] "if"                                 "inherits"                          
#> [123] "INSTALL_opts"                       "interactive"                       
#> [125] "inuse"                              "invisible"                         
#> [127] "is.character"                       "is.list"                           
#> [129] "is.logical"                         "is.na"                             
#> [131] "is.null"                            "isTRUE"                            
#> [133] "keep_outputs"                       "later"                             
#> [135] "length"                             "lib"                               
#> [137] "libpath"                            "libs_only"                         
#> [139] "libs_used"                          "list"                              
#> [141] "make.packages.html"                 "match"                             
#> [143] "message"                            "method"                            
#> [145] "mfile"                              "missing"                           
#> [147] "msg"                                "na"                                
#> [149] "names"                              "Ncpus"                             
#> [151] "needsCmp"                           "next"                              
#> [153] "ngettext"                           "nonlocalrepos"                     
#> [155] "normalizePath"                      "nrow"                              
#> [157] "NROW"                               "nzchar"                            
#> [159] "ok"                                 "OK"                                
#> [161] "oldrlibs"                           "on.exit"                           
#> [163] "OS.type"                            "out"                               
#> [165] "outdir"                             "outfile"                           
#> [167] "outfiles"                           "output"                            
#> [169] "p"                                  "p0"                                
#> [171] "paste"                              "paste0"                            
#> [173] "path.expand"                        "path.sep"                          
#> [175] "pkg"                                "pkgnames"                          
#> [177] "pkgs"                               "pkgType"                           
#> [179] "platform"                           "print"                             
#> [181] "quiet"                              "R.home"                            
#> [183] "R.version"                          "readLines"                         
#> [185] "repos"                              "res"                               
#> [187] "return"                             "row.names"                         
#> [189] "rownames"                           "s2"                                
#> [191] "search"                             "select.list"                       
#> [193] "seq_len"                            "setdiff"                           
#> [195] "setwd"                              "shortPathName"                     
#> [197] "shQuote"                            "sort"                              
#> [199] "sort.list"                          "sprintf"                           
#> [201] "sQuote"                             "srcOnly"                           
#> [203] "srcpkgs"                            "srcvers"                           
#> [205] "startsWith"                         "status"                            
#> [207] "stderr"                             "stop"                              
#> [209] "strsplit"                           "strwrap"                           
#> [211] "sub"                                "sum"                               
#> [213] "suppressWarnings"                   "Sys.getenv"                        
#> [215] "Sys.getpid"                         "Sys.setenv"                        
#> [217] "Sys.which"                          "system"                            
#> [219] "system2"                            "tcltk"                             
#> [221] "tempdir"                            "this"                              
#> [223] "timeout"                            "tlim"                              
#> [225] "tlim_cmd"                           "tmpd"                              
#> [227] "tmpd2"                              "tools"                             
#> [229] "try"                                "tryCatch"                          
#> [231] "tss"                                "type"                              
#> [233] "type2"                              "unique"                            
#> [235] "unlink"                             "unlist"                            
#> [237] "update"                             "upkgs"                             
#> [239] "urls"                               "userdir"                           
#> [241] "valid_package_version"              "verbose"                           
#> [243] "warning"                            "web"                               
#> [245] "writeLines"                        

# Does not find names of arguments
find_symbols(quote(foo(a = A)))

#> [1] "A"   "foo"

#> [1] "A"   "foo"

find_symbols(quote(function(x = X) NULL))

#> [1] "function" "X"       

#> [1] "function" "X"