Given a package namespace or an environment, find all internal symbols used by each function

find_internal_symbols() finds all the objects in the given package namespace (or environment), then, for each function in the environment, it finds all symbols in the function (by calling find_symbols(), and then it filters the symbols so that only symbols referring to objects in the environment/namespace remain.

find_internal_deps() does the same, and then for each function, it considers each (internal) symbol as a dependency and recursively finds all dependencies. So for a given function in the environment, find_internal_deps() finds the internal dependencies needed for that function.

find_internal_symbols(env = "staticimports")

find_internal_deps(env = "staticimports")

Arguments

env

A string naming a package, or an environment.

Value

A named list, where the name of each element is the name of each object in env, and the value is a character vector of strings, where each string is the name of an object in the environment.

Examples

# By default, find symbols internal to the staticimports package
find_internal_symbols()
#> $import
#> [1] "find_staticimports" "import"             "import_objs"       
#> 
#> $find_staticimports
#> [1] "map_chr"
#> 
#> $walk
#> character(0)
#> 
#> $vec_accumulator
#> character(0)
#> 
#> $find_internal_deps_one_impl
#> [1] "find_internal_deps_one_impl"
#> 
#> $process_source_text_one
#> [1] "extract_object_name" "is_parseable"        "vec_accumulator"    
#> 
#> $extract_object_name
#> character(0)
#> 
#> $find_symbols
#> [1] "find_symbols_impl"
#> 
#> $find_internal_deps_one
#> [1] "find_internal_deps_one_impl"
#> 
#> $process_source_texts
#> [1] "process_source_text_one"
#> 
#> $map_chr
#> character(0)
#> 
#> $import_objs
#> [1] "cat0"                 "find_internal_deps"   "is_string"           
#> [4] "process_source_texts"
#> 
#> $find_symbols_impl
#> [1] "find_symbols_impl" "walk"             
#> 
#> $find_internal_symbols
#> [1] "find_symbols" "is_string"   
#> 
#> $is_string
#> character(0)
#> 
#> $find_internal_deps
#> [1] "find_internal_deps_one" "find_internal_symbols"  "is_string"             
#> 
#> $cat0
#> character(0)
#> 
#> $is_parseable
#> character(0)
#> 

# Find all symbols internal to the utils package
find_internal_symbols("utils")
#> $C_sethash_Ext
#> character(0)
#> 
#> $relist.default
#> character(0)
#> 
#> $format.hashtab
#> [1] "numhash" "typhash"
#> 
#> $`[[<-.hashtab`
#> [1] "sethash"
#> 
#> $aspell_control_R_Rd_files
#> character(0)
#> 
#> $stack
#> character(0)
#> 
#> $getS3method
#> [1] ".defenv_for_S3_registry" "findGeneric"            
#> [3] "getKnownS3generics"      "methods"                
#> 
#> $C_menu
#> character(0)
#> 
#> $regquote
#> character(0)
#> 
#> $debugger
#> [1] "C_menu"
#> 
#> $available.packages
#> [1] "URLencode"                          "available_packages_filters_db"     
#> [3] "available_packages_filters_default" "contrib.url"                       
#> [5] "download.file"                      "timestamp"                         
#> 
#> $toLatex.sessionInfo
#> [1] ".RNGdefaults"  "toLatexPDlist"
#> 
#> $registerNames
#> character(0)
#> 
#> $removeSource
#> character(0)
#> 
#> $fuzzyApropos
#> [1] "findFuzzyMatches"
#> 
#> $tail.default
#> [1] "checkHT"    "tail.array"
#> 
#> $printhsearchInternal
#> [1] "browseURL" "help"     
#> 
#> $expr2token
#> character(0)
#> 
#> $methods
#> [1] ".MethodsFunction" ".S3methods"       "methods"         
#> 
#> $personList
#> character(0)
#> 
#> $aspell_package_vignettes
#> [1] ".aspell_package_defaults"         "aspell"                          
#> [3] "aspell_control_package_vignettes" "aspell_find_dictionaries"        
#> [5] "aspell_find_program"             
#> 
#> $packageStatus
#> [1] "available.packages" "contrib.url"        "installed.packages"
#> 
#> $glob2rx
#> character(0)
#> 
#> $makeUserAgent
#> character(0)
#> 
#> $object.size
#> [1] "C_objectSize"
#> 
#> $`$.bibentry`
#> [1] "bibentry_attribute_names"
#> 
#> $count.fields
#> [1] "C_countfields"
#> 
#> $aspell_control_package_vignettes
#> character(0)
#> 
#> $read.socket
#> [1] "C_sockread"
#> 
#> $str2logical
#> character(0)
#> 
#> $vi
#> [1] "edit.default"
#> 
#> $RtangleRuncode
#> [1] ".RtangleCodeLabel"          ".SweaveValidFilenameRegexp"
#> [3] "RweaveChunkPrefix"          "SweaveHooks"               
#> 
#> $as.relistable
#> character(0)
#> 
#> $str.hashtab
#> [1] "numhash" "typhash"
#> 
#> $C_processevents
#> character(0)
#> 
#> $unstack
#> character(0)
#> 
#> $C_sockclose
#> character(0)
#> 
#> $getSrcDirectory
#> [1] "getSrcFilename"
#> 
#> $read.delim2
#> [1] "read.table"
#> 
#> $loadedPackageCompletions
#> [1] ".CompletionEnv"   "findExactMatches" "makeRegexpSafe"   "rc.getOption"    
#> 
#> $offline_help_helper
#> character(0)
#> 
#> $check_for_XQuartz
#> character(0)
#> 
#> $as.environment.hashtab
#> character(0)
#> 
#> $dataentry
#> [1] "C_dataentry"       "check_for_XQuartz" "data"             
#> 
#> $specialOpLocs
#> [1] "specialOps"   "tail.default"
#> 
#> $remhash
#> [1] "C_remhash_Ext"
#> 
#> $write.table
#> [1] "C_writetable"
#> 
#> $promptData
#> [1] "str"
#> 
#> $prompt.default
#> [1] "promptData"
#> 
#> $prompt
#> character(0)
#> 
#> $str.Date
#> [1] "str"         "str.default"
#> 
#> $toBibtex.bibentry
#> [1] "head"
#> 
#> $as.bibentry.bibentry
#> character(0)
#> 
#> $print.packageStatus
#> character(0)
#> 
#> $toBibtex
#> character(0)
#> 
#> $aspell_R_manuals
#> [1] "aspell"                   "aspell_control_R_manuals"
#> [3] "aspell_dictionaries_R"    "aspell_find_program"     
#> 
#> $shorten.to.string
#> character(0)
#> 
#> $available_packages_filters_default
#> character(0)
#> 
#> $write.csv2
#> [1] "write.table"
#> 
#> $relist.factor
#> character(0)
#> 
#> $download.packages
#> [1] "available.packages" "contrib.url"        "download.file"     
#> [4] "resolvePkgType"    
#> 
#> $install.packages
#>  [1] ".R_LIBS"                            ".find_install_order"               
#>  [3] ".install.macbinary"                 ".make_dependency_list"             
#>  [5] "askYesNo"                           "available.packages"                
#>  [7] "available_packages_filters_db"      "available_packages_filters_default"
#>  [9] "contrib.url"                        "download.file"                     
#> [11] "download.packages"                  "flush.console"                     
#> [13] "getDependencies"                    "make.packages.html"                
#> [15] "select.list"                       
#> 
#> $conformToProto
#> [1] "methods"
#> 
#> $help
#> [1] "browseURL"    "index.search"
#> 
#> $head.ftable
#> [1] "head.matrix"
#> 
#> $print.person
#> character(0)
#> 
#> $citation
#>  [1] ".citation"                ".read_authors_at_R_field"
#>  [3] "as.personList"            "bibentry"                
#>  [5] "citation"                 "file_test"               
#>  [7] "head"                     "packageDescription"      
#>  [9] "readCitationFile"         "tail"                    
#> 
#> $zip
#> [1] "zip"
#> 
#> $savehistory
#> [1] "C_savehistory"
#> 
#> $readCitationFile
#> [1] ".citation"             ".is_not_nonempty_text"
#> 
#> $aspell_dictionaries_R
#> character(0)
#> 
#> $View
#> [1] "C_dataviewer"      "check_for_XQuartz"
#> 
#> $findFuzzyMatches
#> [1] "findExactMatches" "rc.getOption"    
#> 
#> $`?`
#> [1] ".helpForCall"  ".tryHelp"      "flush.console" "help"         
#> [5] "help.search"   "topicName"    
#> 
#> $str.POSIXt
#> [1] "str"         "str.default"
#> 
#> $getParseText
#> [1] "substr_with_tabs"
#> 
#> $C_Rprof
#> character(0)
#> 
#> $rc.getOption
#> [1] ".CompletionEnv"
#> 
#> $summaryRprof
#> [1] "Rprof_memory_summary" "tail"                
#> 
#> $old.packages
#> [1] "available.packages" "contrib.url"        "installed.packages"
#> 
#> $getTxtProgressBar
#> character(0)
#> 
#> $read.csv2
#> [1] "read.table"
#> 
#> $xemacs
#> [1] "edit.default"
#> 
#> $specialFunctionArgs
#> [1] ".CompletionEnv"     "data"               "findMatches"       
#> [4] "installed.packages" "makeRegexpSafe"    
#> 
#> $substr_with_tabs
#> character(0)
#> 
#> $aspell_control_R_manuals
#> character(0)
#> 
#> $tail.matrix
#> [1] "checkHT"
#> 
#> $read.delim
#> [1] "read.table"
#> 
#> $`[.roman`
#> character(0)
#> 
#> $citFooter
#> character(0)
#> 
#> $bibentry_format_styles
#> character(0)
#> 
#> $attachedPackageCompletions
#> [1] ".CompletionEnv"   "findExactMatches" "makeRegexpSafe"   "rc.getOption"    
#> 
#> $aspell_package_C_files
#> [1] ".aspell_package_defaults" "aspell"                  
#> [3] "aspell_find_dictionaries" "aspell_find_program"     
#> [5] "file_test"               
#> 
#> $checkHT
#> character(0)
#> 
#> $breakRE
#> character(0)
#> 
#> $read.fortran
#> [1] "read.fwf"
#> 
#> $as.person
#> character(0)
#> 
#> $find_files_in_directories
#> [1] "file_test"
#> 
#> $findGeneric
#> [1] "methods"
#> 
#> $capture.output
#> character(0)
#> 
#> $getSrcfile
#> [1] "getSrcref"
#> 
#> $close.txtProgressBar
#> character(0)
#> 
#> $read.DIF
#> [1] "data"         "methods"      "type.convert"
#> 
#> $print.packageIQR
#> character(0)
#> 
#> $write.etags
#> [1] "shorten.to.string"
#> 
#> $C_typeconvert
#> character(0)
#> 
#> $RweaveLatexSetup
#> [1] "RweaveLatexOptions"
#> 
#> $c.bibentry
#> [1] ".bibentry"
#> 
#> $vignette
#> [1] "file_test"
#> 
#> $getSrcref
#> [1] "getSrcref" "methods"  
#> 
#> $isBasePkg
#> [1] "packageDescription"
#> 
#> $blank_out_ignores_in_lines
#> [1] "blank_out_regexp_matches"
#> 
#> $matchAvailableTopics
#> [1] ".CompletionEnv" "findMatches"    "makeRegexpSafe"
#> 
#> $hsearch_db_types
#> character(0)
#> 
#> $debugcall
#> [1] ".debugcall"
#> 
#> $`[.person`
#> character(0)
#> 
#> $index.search
#> character(0)
#> 
#> $nsl
#> [1] "C_nsl"
#> 
#> $RweaveLatex
#> [1] "RweaveLatexFinish"   "RweaveLatexOptions"  "RweaveLatexRuncode" 
#> [4] "RweaveLatexSetup"    "RweaveLatexWritedoc"
#> 
#> $C_writetable
#> character(0)
#> 
#> $package.skeleton
#> [1] ".fixPackageFileNames" "methods"              "prompt"              
#> [4] "promptPackage"       
#> 
#> $print.MethodsFunction
#> character(0)
#> 
#> $MARC_relator_db_codes_used_with_R
#> character(0)
#> 
#> $normalCompletions
#> [1] ".CompletionEnv"           "apropos"                 
#> [3] "fuzzyApropos"             "keywordCompletions"      
#> [5] "loadedPackageCompletions" "makeRegexpSafe"          
#> [7] "rc.getOption"            
#> 
#> $getCRANmirrors
#> [1] ".getMirrors"
#> 
#> $C_ishashtab_Ext
#> character(0)
#> 
#> $de.setup
#> character(0)
#> 
#> $head.matrix
#> [1] "checkHT"
#> 
#> $RtangleFinish
#> character(0)
#> 
#> $Rprof
#> [1] "C_Rprof"
#> 
#> $tail.function
#> [1] "checkHT" "tail"   
#> 
#> $read.fwf
#> [1] "read.table"
#> 
#> $C_download
#> character(0)
#> 
#> $promptImport
#> character(0)
#> 
#> $print.vignette
#> [1] "browseURL"
#> 
#> $filter_packages_by_depends_predicates
#> character(0)
#> 
#> $alarm
#> [1] "flush.console"
#> 
#> $Ops.roman
#> [1] ".as.roman" "as.roman" 
#> 
#> $summary.packageStatus
#> character(0)
#> 
#> $relist.list
#> [1] "relist"
#> 
#> $tail.ftable
#> [1] "tail.matrix"
#> 
#> $isInsideQuotes
#> [1] ".CompletionEnv" "head.default"  
#> 
#> $specialOps
#> character(0)
#> 
#> $length.hashtab
#> [1] "numhash"
#> 
#> $print.news_db
#> [1] "browseURL"
#> 
#> $type.convert.list
#> [1] "type.convert"
#> 
#> $format.aspell
#> character(0)
#> 
#> $C_dataentry
#> character(0)
#> 
#> $read.table
#> [1] "C_readtablehead" "data"            "methods"         "type.convert"   
#> 
#> $Rtangle
#> [1] "RtangleFinish"      "RtangleRuncode"     "RtangleSetup"      
#> [4] "RtangleWritedoc"    "RweaveLatexOptions"
#> 
#> $packageDescription
#> character(0)
#> 
#> $aspell_package_description
#> [1] ".aspell_package_defaults" "aspell"                  
#> [3] "aspell_find_dictionaries" "aspell_find_program"     
#> 
#> $RShowDoc
#> [1] "URLencode" "browseURL" "file_test"
#> 
#> $print.Bibtex
#> character(0)
#> 
#> $rep.roman
#> character(0)
#> 
#> $as.person.default
#> [1] ".is_not_nonempty_text" "person"               
#> 
#> $timestamp
#> [1] "C_addhistory"
#> 
#> $findLineNum
#> [1] "fnLineNum" "lsf.str"   "methods"  
#> 
#> $print.summary.packageStatus
#> character(0)
#> 
#> $bug.report
#> [1] "browseURL"          "bug.report.info"    "create.post"       
#> [4] "flush.console"      "packageDescription"
#> 
#> $strcapture
#> [1] "conformToProto" "str"           
#> 
#> $print.bibentry
#> [1] ".bibentry_match_format_style"
#> 
#> $URLdecode
#> character(0)
#> 
#> $unlist.relistable
#> character(0)
#> 
#> $rc.options
#> [1] ".CompletionEnv" "modifyList"    
#> 
#> $file.edit
#> [1] "C_fileedit"
#> 
#> $make_sysdata_rda
#> [1] "MARC_relator_db"                   "MARC_relator_db_codes_used_with_R"
#> 
#> $Stangle
#> [1] "Rtangle" "Sweave" 
#> 
#> $untar
#> [1] "tar"    "untar2"
#> 
#> $C_selectlist
#> character(0)
#> 
#> $print.getAnywhere
#> character(0)
#> 
#> $C_socklisten
#> character(0)
#> 
#> $getSrcLocation
#> [1] "getSrcLocation" "getSrcref"     
#> 
#> $findCRANmirror
#> character(0)
#> 
#> $write.csv
#> [1] "write.table"
#> 
#> $as.character.person
#> character(0)
#> 
#> $txtProgressBar
#> [1] "flush.console"
#> 
#> $bug.report.info
#> character(0)
#> 
#> $isS3method
#> [1] ".defenv_for_S3_registry" "findGeneric"            
#> [3] "getKnownS3generics"      "isS3method"             
#> [5] "methods"                
#> 
#> $isS3stdGeneric
#> [1] "methods"
#> 
#> $fileSnapshot
#> [1] "timestamp"
#> 
#> $help.search
#> [1] "apropos"           "flush.console"     "hsearch_db"       
#> [4] "hsearch_db_fields" "hsearch_db_types" 
#> 
#> $toBibtex.person
#> character(0)
#> 
#> $aspell_R_C_files
#> [1] "aspell"                    "aspell_control_R_Rd_files"
#> [3] "aspell_dictionaries_R"     "aspell_find_program"      
#> [5] "file_test"                
#> 
#> $getRcode
#> character(0)
#> 
#> $packageName
#> character(0)
#> 
#> $browseVignettes
#> character(0)
#> 
#> $combn
#> character(0)
#> 
#> $findMatches
#> [1] ".CompletionEnv"   "findExactMatches" "findFuzzyMatches"
#> 
#> $bibentry_list_attribute_names
#> character(0)
#> 
#> $undebugcall
#> [1] ".debugcall"
#> 
#> $as.roman
#> [1] ".as.roman"
#> 
#> $contrib.url
#> [1] "chooseCRANmirror" "flush.console"    "resolvePkgType"  
#> 
#> $Summary.roman
#> [1] "as.roman"
#> 
#> $format.bibentry
#> [1] ".bibentry"                    ".bibentry_expand_crossrefs"  
#> [3] ".bibentry_match_format_style" ".format_bibentry_as_R_code"  
#> [5] "toBibtex"                    
#> 
#> $URLencode
#> character(0)
#> 
#> $SweaveParseOptions
#> character(0)
#> 
#> $print.changedFiles
#> character(0)
#> 
#> $`$.person`
#> [1] "head" "tail"
#> 
#> $format.news_db
#> character(0)
#> 
#> $defaultUserAgent
#> character(0)
#> 
#> $SweaveGetSyntax
#> [1] "SweaveSyntaxNoweb" "apropos"          
#> 
#> $C_Rprofmem
#> character(0)
#> 
#> $bibentry
#> [1] ".bibentry"                 ".bibentry_check_bibentry1"
#> [3] ".is_not_nonempty_text"     ".listify"                 
#> [5] "as.person"                
#> 
#> $ls.str
#> character(0)
#> 
#> $browseEnv
#> [1] "wsbrowser"
#> 
#> $specialCompletions
#> [1] "helpCompletions"            "specialOpCompletionsHelper"
#> 
#> $RweaveLatexRuncode
#> [1] ".SweaveValidFilenameRegexp" "RweaveChunkPrefix"         
#> [3] "RweaveTryStop"              "SweaveHooks"               
#> 
#> $aspell_R_vignettes
#> [1] "aspell"                     "aspell_control_R_vignettes"
#> [3] "aspell_dictionaries_R"      "aspell_find_program"       
#> 
#> $download.file
#> [1] "C_download"    "flush.console"
#> 
#> $edit.default
#> [1] "C_edit"
#> 
#> $lsf.str
#> [1] "ls.str"
#> 
#> $checkCRAN
#> [1] "available.packages" "contrib.url"        "getCRANmirrors"    
#> 
#> $hashtab
#> [1] "C_hashtab_Ext"
#> 
#> $getDependencies
#> [1] ".clean_up_dependencies2" "flush.console"          
#> [3] "installed.packages"      "isBasePkg"              
#> 
#> $C_sockread
#> character(0)
#> 
#> $aspell_package
#> [1] "aspell_package_C_files"     "aspell_package_R_files"    
#> [3] "aspell_package_Rd_files"    "aspell_package_description"
#> [5] "aspell_package_vignettes"  
#> 
#> $head.array
#> [1] "checkHT"
#> 
#> $upgrade
#> character(0)
#> 
#> $news
#> character(0)
#> 
#> $print.browseVignettes
#> [1] "browseURL"
#> 
#> $strOptions
#> character(0)
#> 
#> $C_gethash_Ext
#> character(0)
#> 
#> $C_readtablehead
#> character(0)
#> 
#> $C_sockconnect
#> character(0)
#> 
#> $print.aspell
#> character(0)
#> 
#> $setBreakpoint
#> [1] "findLineNum"
#> 
#> $packageDate
#> [1] "asDateBuilt"        "packageDescription"
#> 
#> $C_dataviewer
#> character(0)
#> 
#> $argNames
#> [1] ".CompletionEnv" ".FunArgEnv"     "argsAnywhere"  
#> 
#> $RweaveChunkPrefix
#> character(0)
#> 
#> $keywordCompletions
#> [1] "findExactMatches" "makeRegexpSafe"  
#> 
#> $mirror2html
#> [1] "getCRANmirrors" "head"          
#> 
#> $page
#> character(0)
#> 
#> $getAnywhere
#> [1] "find"        "getS3method"
#> 
#> $osVersion
#> character(0)
#> 
#> $format.citation
#> [1] "format.bibentry"
#> 
#> $apropos
#> [1] ".dot_internals"
#> 
#> $print.help_files_with_topic
#> [1] ".getHelpFile"                ".show_help_on_topic_offline"
#> [3] "URLencode"                   "browseURL"                  
#> [5] "menu"                       
#> 
#> $type.convert.data.frame
#> [1] "type.convert"
#> 
#> $de
#> [1] "dataentry"  "de.ncols"   "de.restore" "de.setup"  
#> 
#> $type.convert.default
#> [1] "C_typeconvert"
#> 
#> $Sweave
#> [1] "RweaveLatex"        "SweaveGetSyntax"    "SweaveParseOptions"
#> [4] "SweaveReadFile"    
#> 
#> $specialOpCompletionsHelper
#> [1] ".CompletionEnv"    ".DollarNames"      "findMatches"      
#> [4] "makeRegexpSafe"    "methods"           "normalCompletions"
#> 
#> $installed.packages
#> [1] ".fixupPkgMat"   ".instPkgFields" ".readPkgDesc"   "C_crc64"       
#> 
#> $format.object_size
#> character(0)
#> 
#> $SweaveSyntConv
#> [1] "SweaveGetSyntax"
#> 
#> $rc.status
#> [1] ".CompletionEnv"
#> 
#> $hsearch_db_keywords
#> [1] ".get_standard_Rd_keywords_with_descriptions"
#> [2] "hsearch_db"                                 
#> 
#> $RweaveLatexOptions
#> character(0)
#> 
#> $available_packages_filters_db
#> character(0)
#> 
#> $c.person
#> character(0)
#> 
#> $packageVersion
#> [1] "packageDescription"
#> 
#> $SweaveHooks
#> [1] "SweaveHooks"
#> 
#> $C_numhash_Ext
#> character(0)
#> 
#> $rtags.file
#> [1] "expr2token"  "write.etags"
#> 
#> $`[.getAnywhere`
#> character(0)
#> 
#> $person
#> [1] ".canonicalize_person_role" ".is_not_nonempty_text"    
#> [3] ".listify"                  ".person"                  
#> 
#> $pico
#> [1] "edit.default"
#> 
#> $citeNatbib
#> [1] "person"
#> 
#> $help.start
#> [1] "browseURL"          "make.packages.html"
#> 
#> $topicName
#> character(0)
#> 
#> $wsbrowser
#> [1] "URLencode" "browseURL"
#> 
#> $functionArgs
#> [1] ".CompletionEnv"      "argNames"            "findMatches"        
#> [4] "makeRegexpSafe"      "methods"             "rc.getOption"       
#> [7] "specialFunctionArgs"
#> 
#> $maphash
#> [1] "C_maphash_Ext"
#> 
#> $getSrcByte
#> character(0)
#> 
#> $print.hashtab
#> character(0)
#> 
#> $url.show
#> [1] "download.file"
#> 
#> $edit
#> character(0)
#> 
#> $make.packages.html
#> [1] "URLencode"          "flush.console"      "packageDescription"
#> 
#> $aspell
#> [1] "aspell_filter_db"                     
#> [2] "aspell_find_dictionaries"             
#> [3] "aspell_find_program"                  
#> [4] "aspell_write_personal_dictionary_file"
#> 
#> $browseURL
#> [1] "URLencode"
#> 
#> $correctFilenameToken
#> [1] ".CompletionEnv" "head.default"  
#> 
#> $sethash
#> [1] "C_sethash_Ext"
#> 
#> $data.entry
#> [1] "de"
#> 
#> $is.hashtab
#> [1] "C_ishashtab_Ext"
#> 
#> $promptPackage
#> character(0)
#> 
#> $memory.size
#> character(0)
#> 
#> $maintainer
#> [1] "packageDescription"
#> 
#> $C_countfields
#> character(0)
#> 
#> $aspell_R_Rd_files
#> [1] "aspell"                    "aspell_control_R_Rd_files"
#> [3] "aspell_dictionaries_R"     "aspell_find_program"      
#> [5] "file_test"                
#> 
#> $dump.frames
#> [1] "limitedLabels"
#> 
#> $format.MethodsFunction
#> character(0)
#> 
#> $blank_out_regexp_matches
#> character(0)
#> 
#> $fnLineNum
#> [1] ".normalizePath" "timestamp"     
#> 
#> $aspell_package_R_files
#> [1] ".aspell_package_defaults" "aspell"                  
#> [3] "aspell_find_dictionaries" "aspell_find_program"     
#> 
#> $C_typhash_Ext
#> character(0)
#> 
#> $C_loadhistory
#> character(0)
#> 
#> $tail.array
#> [1] "checkHT"
#> 
#> $head.function
#> [1] "checkHT" "head"   
#> 
#> $as.bibentry.citation
#> character(0)
#> 
#> $strextract
#> character(0)
#> 
#> $edit.vignette
#> [1] "file.edit" "getRcode" 
#> 
#> $strslice
#> [1] "conformToProto" "str"           
#> 
#> $`[[.person`
#> character(0)
#> 
#> $`[[.hashtab`
#> [1] "gethash"
#> 
#> $hsearch_db
#> [1] ".get_standard_Rd_keywords_with_descriptions"
#> [2] ".hsearch_db"                                
#> [3] "flush.console"                              
#> [4] "hsearch_db_types"                           
#> [5] "merge_demo_index"                           
#> [6] "merge_vignette_index"                       
#> 
#> $write.ctags
#> character(0)
#> 
#> $hsearch_db_fields
#> character(0)
#> 
#> $find
#> [1] ".dot_internals"
#> 
#> $str.default
#> [1] "%||%"           "capture.output" "methods"        "modifyList"    
#> [5] "str"            "str.default"    "strOptions"    
#> 
#> $C_addhistory
#> character(0)
#> 
#> $update.packages
#> [1] "askYesNo"           "available.packages" "contrib.url"       
#> [4] "install.packages"   "old.packages"       "select.list"       
#> [7] "simplifyRepos"     
#> 
#> $print.Latex
#> character(0)
#> 
#> $create.post
#> [1] "URLencode"     "askYesNo"      "file.edit"     "flush.console"
#> 
#> $unstack.data.frame
#> character(0)
#> 
#> $aspell_package_Rd_files
#> [1] ".aspell_package_defaults" "aspell"                  
#> [3] "aspell_find_dictionaries"
#> 
#> $subset.news_db
#> [1] "news"
#> 
#> $aspell_control_R_vignettes
#> character(0)
#> 
#> $memory.limit
#> character(0)
#> 
#> $print.findLineNumResult
#> character(0)
#> 
#> $recover
#> [1] "dump.frames"   "limitedLabels" "menu"          "recover"      
#> 
#> $example
#> [1] ".getHelpFile" "browseURL"    "index.search" "str2logical" 
#> 
#> $chooseCRANmirror
#> [1] ".chooseMirror"  "getCRANmirrors"
#> 
#> $chooseBioCmirror
#> [1] ".chooseMirror" ".getMirrors"  
#> 
#> $write.socket
#> [1] "C_sockwrite"
#> 
#> $formatUL
#> [1] ".format_rl_table"
#> 
#> $aspell_inspect_context
#> character(0)
#> 
#> $select.list
#> [1] "C_selectlist" "menu"        
#> 
#> $SweaveSyntaxNoweb
#> character(0)
#> 
#> $limitedLabels
#> character(0)
#> 
#> $print.sessionInfo
#> [1] ".RNGdefaults"
#> 
#> $print.object_size
#> [1] "format.object_size"
#> 
#> $Rprof_memory_summary
#> character(0)
#> 
#> $relist.matrix
#> [1] "relist"
#> 
#> $`%||%`
#> character(0)
#> 
#> $as.bibentry
#> character(0)
#> 
#> $getRcode.vignette
#> character(0)
#> 
#> $setRepositories
#> [1] "select.list"
#> 
#> $makeRweaveLatexCodeRunner
#> [1] ".SweaveValidFilenameRegexp" "RweaveChunkPrefix"         
#> [3] "RweaveEvalWithOpt"          "RweaveTryStop"             
#> [5] "SweaveHooks"               
#> 
#> $globalVariables
#> [1] "registerNames"
#> 
#> $SweaveSyntaxLatex
#> character(0)
#> 
#> $help.request
#>  [1] ".instPkgFields"  "RSiteSearch"     "askYesNo"        "browseURL"      
#>  [5] "bug.report.info" "create.post"     "old.packages"    "page"           
#>  [9] "sessionInfo"     "update.packages"
#> 
#> $unzip
#> [1] "C_unzip"    "read.table" "unzip"     
#> 
#> $RweaveTryStop
#> character(0)
#> 
#> $fixInNamespace
#> [1] "assignInNamespace" "edit"             
#> 
#> $read.csv
#> [1] "read.table"
#> 
#> $getKnownS3generics
#> character(0)
#> 
#> $is.relistable
#> character(0)
#> 
#> $edit.data.frame
#> [1] "C_dataentry"       "check_for_XQuartz" "edit.default"     
#> 
#> $rep.bibentry
#> character(0)
#> 
#> $print.ls_str
#> [1] "str"
#> 
#> $asDateBuilt
#> character(0)
#> 
#> $gethash
#> [1] "C_gethash_Ext"
#> 
#> $`$<-.bibentry`
#> [1] ".bibentry"                 ".bibentry_check_bibentry1"
#> [3] ".listify"                  "as.person"                
#> [5] "bibentry_attribute_names" 
#> 
#> $unique.bibentry
#> character(0)
#> 
#> $setTxtProgressBar
#> character(0)
#> 
#> $C_nsl
#> character(0)
#> 
#> $aspell_filter_db
#> character(0)
#> 
#> $setIsFirstArg
#> [1] ".CompletionEnv"
#> 
#> $upgrade.packageStatus
#> [1] "askYesNo"         "install.packages" "write.table"     
#> 
#> $RSiteSearch
#> [1] "URLencode" "browseURL"
#> 
#> $unstack.default
#> character(0)
#> 
#> $C_flushconsole
#> character(0)
#> 
#> $aspell_write_personal_dictionary_file
#> [1] "aspell_find_program"
#> 
#> $aspell_R_R_files
#> [1] "aspell"                    "aspell_control_R_Rd_files"
#> [3] "aspell_dictionaries_R"     "aspell_find_program"      
#> 
#> $update.packageStatus
#> [1] "packageStatus"
#> 
#> $C_edit
#> character(0)
#> 
#> $getFromNamespace
#> character(0)
#> 
#> $history
#> [1] "savehistory"
#> 
#> $format.person
#> [1] ".expand_ORCID_identifier" ".format_person_as_R_code"
#> 
#> $aspell_package_pot_files
#> [1] "aspell"              "aspell_find_program"
#> 
#> $as.personList.default
#> [1] "as.person"
#> 
#> $head
#> character(0)
#> 
#> $fix
#> [1] "edit"
#> 
#> $C_objectSize
#> character(0)
#> 
#> $prompt.data.frame
#> [1] "promptData"
#> 
#> $hasName
#> character(0)
#> 
#> $askYesNo
#> character(0)
#> 
#> $helpCompletions
#> [1] ".CompletionEnv"       "matchAvailableTopics" "normalCompletions"   
#> 
#> $process.events
#> [1] "C_processevents"
#> 
#> $C_crc64
#> character(0)
#> 
#> $localeToCharset
#> character(0)
#> 
#> $as.personList
#> character(0)
#> 
#> $as.character.roman
#> [1] ".numeric2roman"
#> 
#> $tar
#> [1] "C_octsize" "tar"      
#> 
#> $citHeader
#> character(0)
#> 
#> $SweaveReadFile
#> [1] "SweaveReadFile"
#> 
#> $charClass
#> [1] "C_charClass"
#> 
#> $RtangleSetup
#> [1] "RweaveLatexOptions"
#> 
#> $de.restore
#> character(0)
#> 
#> $`[.news_db`
#> character(0)
#> 
#> $C_savehistory
#> character(0)
#> 
#> $`$<-.citation`
#> [1] ".citation"
#> 
#> $get_parse_data_for_message_strings
#> [1] "getParseData" "getParseText"
#> 
#> $MARC_relator_db
#> character(0)
#> 
#> $type.convert
#> character(0)
#> 
#> $merge_vignette_index
#> character(0)
#> 
#> $typhash
#> [1] "C_typhash_Ext"
#> 
#> $fileCompletionPreferred
#> [1] ".CompletionEnv" "head.default"  
#> 
#> $print.citation
#> [1] "print.bibentry"
#> 
#> $sessionInfo
#> [1] "osVersion"          "packageDescription"
#> 
#> $cite
#> [1] "cite"       "citeNatbib"
#> 
#> $as.personList.person
#> character(0)
#> 
#> $getIsFirstArg
#> [1] ".CompletionEnv"
#> 
#> $C_maphash_Ext
#> character(0)
#> 
#> $file_test
#> character(0)
#> 
#> $summary.aspell
#> character(0)
#> 
#> $print.aspell_inspect_context
#> character(0)
#> 
#> $menu
#> [1] "C_menu"      "select.list"
#> 
#> $str.data.frame
#> character(0)
#> 
#> $bibentry_attribute_names
#> character(0)
#> 
#> $print.hsearch_db
#> character(0)
#> 
#> $rc.settings
#> [1] ".CompletionEnv" "data"           "help"          
#> 
#> $tail
#> character(0)
#> 
#> $close.socket
#> [1] "C_sockclose"
#> 
#> $makeRegexpSafe
#> character(0)
#> 
#> $simplifyRepos
#> [1] "contrib.url" "tail"       
#> 
#> $C_sockwrite
#> character(0)
#> 
#> $getParseData
#> [1] "data"         "getParseText" "getSrcfile"  
#> 
#> $C_sockopen
#> character(0)
#> 
#> $str
#> character(0)
#> 
#> $format.roman
#> [1] "as.character.roman"
#> 
#> $aregexec
#> [1] "aregexec"
#> 
#> $demo
#> [1] "browseURL"   "file_test"   "str2logical"
#> 
#> $changedFiles
#> [1] "fileSnapshot" "file_test"    "timestamp"   
#> 
#> $C_unzip
#> character(0)
#> 
#> $adist
#> [1] "adist"
#> 
#> $toLatexPDlist
#> character(0)
#> 
#> $untar2
#> character(0)
#> 
#> $assignInMyNamespace
#> [1] ".defenv_for_S3_registry" "methods"                
#> 
#> $print.roman
#> [1] "as.character.roman"
#> 
#> $inFunction
#> [1] ".CompletionEnv" "breakRE"        "setIsFirstArg"  "tail.default"  
#> 
#> $print.packageDescription
#> character(0)
#> 
#> $findExactMatches
#> character(0)
#> 
#> $print.hsearch
#> [1] "printhsearchInternal"
#> 
#> $suppressForeignCheck
#> [1] "registerNames"
#> 
#> $RweaveLatexFinish
#> character(0)
#> 
#> $RweaveEvalWithOpt
#> [1] "methods"
#> 
#> $C_octsize
#> character(0)
#> 
#> $C_remhash_Ext
#> character(0)
#> 
#> $warnErrList
#> [1] "capture.output"
#> 
#> $modifyList
#> [1] "modifyList"
#> 
#> $`[[.bibentry`
#> [1] ".bibentry_get_key"             "bibentry_list_attribute_names"
#> 
#> $new.packages
#> [1] "available.packages" "contrib.url"        "install.packages"  
#> [4] "installed.packages" "select.list"       
#> 
#> $flush.console
#> [1] "C_flushconsole"
#> 
#> $argsAnywhere
#> [1] "getAnywhere"
#> 
#> $C_charClass
#> character(0)
#> 
#> $emacs
#> [1] "edit.default"
#> 
#> $print.fileSnapshot
#> [1] "timestamp"
#> 
#> $`$<-.person`
#> [1] ".canonicalize_person_role" ".is_not_nonempty_text"    
#> [3] ".listify"                 
#> 
#> $stack.data.frame
#> character(0)
#> 
#> $aspell_find_dictionaries
#> [1] "file_test"                 "find_files_in_directories"
#> 
#> $xedit
#> [1] "edit.default"
#> 
#> $loadhistory
#> [1] "C_loadhistory"
#> 
#> $remove.packages
#> [1] "isBasePkg"          "make.packages.html"
#> 
#> $clrhash
#> [1] "C_clrhash_Ext"
#> 
#> $getSrcFilename
#> [1] "getSrcFilename" "getSrcref"     
#> 
#> $C_clrhash_Ext
#> character(0)
#> 
#> $assignInNamespace
#> [1] ".defenv_for_S3_registry" "methods"                
#> 
#> $stack.default
#> character(0)
#> 
#> $compareVersion
#> character(0)
#> 
#> $rtags
#> [1] "rtags.file"  "write.ctags" "write.etags"
#> 
#> $RweaveLatexWritedoc
#> [1] "RweaveChunkPrefix"  "RweaveLatexOptions" "SweaveParseOptions"
#> 
#> $C_hashtab_Ext
#> character(0)
#> 
#> $make.socket
#> [1] "C_sockclose"   "C_sockconnect" "C_socklisten"  "C_sockopen"   
#> 
#> $toLatex
#> character(0)
#> 
#> $head.default
#> [1] "checkHT"    "head.array"
#> 
#> $Rprofmem
#> [1] "C_Rprofmem"
#> 
#> $resolvePkgType
#> character(0)
#> 
#> $C_fileedit
#> character(0)
#> 
#> $formatOL
#> [1] ".format_rl_table" "as.roman"        
#> 
#> $edit.matrix
#> [1] "C_dataentry"       "check_for_XQuartz" "edit.default"     
#> 
#> $citEntry
#> [1] "bibentry"
#> 
#> $data
#> [1] "C_unzip"     "file_test"   "packageName" "read.table" 
#> 
#> $de.ncols
#> character(0)
#> 
#> $sort.bibentry
#> character(0)
#> 
#> $hsearch_db_concepts
#> [1] "hsearch_db"
#> 
#> $print.socket
#> character(0)
#> 
#> $merge_demo_index
#> character(0)
#> 
#> $numhash
#> [1] "C_numhash_Ext"
#> 
#> $aspell_find_program
#> character(0)
#> 
#> $`[.bibentry`
#> [1] ".bibentry_get_key"             "bibentry_list_attribute_names"
#> 
#> $relist
#> character(0)
#> 
#> $fileCompletions
#> [1] "correctFilenameToken"
#> 
#> $RtangleWritedoc
#> [1] "RweaveLatexOptions" "SweaveParseOptions"
#>