Find all symbols used in a function or language object.

Find all symbols used in a function or language object.

find_symbols(x)

Arguments

x
A language object or function.

Value

A character vector of the symbols used in the object.

Examples

# Find all symbols used in the lm function
find_symbols(utils::install.packages)
#>   [1] "::"                                 ":::"                               
#>   [3] "!"                                  "!="                                
#>   [5] "..."                                ".find_install_order"               
#>   [7] ".install.macbinary"                 ".install.winbinary"                
#>   [9] ".libPaths"                          ".Library"                          
#>  [11] ".make_dependency_list"              ".Platform"                         
#>  [13] ".R_LIBS"                            ".standard_regexps"                 
#>  [15] ".TkUp"                              "("                                 
#>  [17] "["                                  "[["                                
#>  [19] "{"                                  "&"                                 
#>  [21] "&&"                                 "%in%"                              
#>  [23] "<"                                  "<-"                                
#>  [25] "=="                                 ">"                                 
#>  [27] "|"                                  "||"                                
#>  [29] "$"                                  "action"                            
#>  [31] "aDL"                                "all"                               
#>  [33] "ans"                                "any"                               
#>  [35] "args"                               "args0"                             
#>  [37] "as.numeric_version"                 "askYesNo"                          
#>  [39] "av"                                 "av1"                               
#>  [41] "av2"                                "available"                         
#>  [43] "available_packages_filters_db"      "available_packages_filters_default"
#>  [45] "available.packages"                 "basename"                          
#>  [47] "bins"                               "binvers"                           
#>  [49] "c"                                  "capabilities"                      
#>  [51] "cat"                                "cbind"                             
#>  [53] "character"                          "clean"                             
#>  [55] "close"                              "cmd"                               
#>  [57] "cmd0"                               "colnames"                          
#>  [59] "config"                             "configure.args"                    
#>  [61] "configure.vars"                     "conn"                              
#>  [63] "contrib.url"                        "contriburl"                        
#>  [65] "contriburl2"                        "cwd"                               
#>  [67] "data.frame"                         "dependencies"                      
#>  [69] "deps"                               "destdir"                           
#>  [71] "destfile"                           "df"                                
#>  [73] "dir"                                "dir.create"                        
#>  [75] "dir.exists"                         "dirname"                           
#>  [77] "dirs"                               "DL"                                
#>  [79] "download.file"                      "download.packages"                 
#>  [81] "e"                                  "endsWith"                          
#>  [83] "env"                                "f"                                 
#>  [85] "f2"                                 "failed"                            
#>  [87] "fil"                                "file"                              
#>  [89] "file.access"                        "file.copy"                         
#>  [91] "file.exists"                        "file.path"                         
#>  [93] "files"                              "filters"                           
#>  [95] "flush.console"                      "fn"                                
#>  [97] "for"                                "found"                             
#>  [99] "foundpkgs"                          "function"                          
#> [101] "get_install_opts"                   "get_package_name"                  
#> [103] "get_timeout"                        "getConfigureArgs"                  
#> [105] "getConfigureVars"                   "getDependencies"                   
#> [107] "getDeps"                            "getOption"                         
#> [109] "gettext"                            "gettextf"                          
#> [111] "getwd"                              "grep"                              
#> [113] "grepl"                              "gsub"                              
#> [115] "GUI"                                "hasArchs"                          
#> [117] "hasSrc"                             "have_spaces"                       
#> [119] "i"                                  "if"                                
#> [121] "inherits"                           "INSTALL_opts"                      
#> [123] "interactive"                        "inuse"                             
#> [125] "invisible"                          "is.character"                      
#> [127] "is.list"                            "is.logical"                        
#> [129] "is.na"                              "is.null"                           
#> [131] "isTRUE"                             "keep_outputs"                      
#> [133] "later"                              "length"                            
#> [135] "lib"                                "libpath"                           
#> [137] "libs_only"                          "libs_used"                         
#> [139] "list"                               "make.packages.html"                
#> [141] "match"                              "message"                           
#> [143] "method"                             "mfile"                             
#> [145] "missing"                            "msg"                               
#> [147] "na"                                 "names"                             
#> [149] "Ncpus"                              "needsCmp"                          
#> [151] "next"                               "ngettext"                          
#> [153] "nonlocalrepos"                      "normalizePath"                     
#> [155] "nrow"                               "NROW"                              
#> [157] "nzchar"                             "ok"                                
#> [159] "OK"                                 "oldrlibs"                          
#> [161] "on.exit"                            "OS.type"                           
#> [163] "out"                                "outdir"                            
#> [165] "outfile"                            "outfiles"                          
#> [167] "output"                             "p"                                 
#> [169] "p0"                                 "paste"                             
#> [171] "paste0"                             "path.expand"                       
#> [173] "path.sep"                           "pkg"                               
#> [175] "pkgnames"                           "pkgs"                              
#> [177] "pkgType"                            "platform"                          
#> [179] "print"                              "quiet"                             
#> [181] "R.home"                             "R.version"                         
#> [183] "readLines"                          "repos"                             
#> [185] "res"                                "return"                            
#> [187] "row.names"                          "rownames"                          
#> [189] "s2"                                 "search"                            
#> [191] "select.list"                        "seq_len"                           
#> [193] "setdiff"                            "setwd"                             
#> [195] "shortPathName"                      "shQuote"                           
#> [197] "sort"                               "sort.list"                         
#> [199] "sprintf"                            "sQuote"                            
#> [201] "srcOnly"                            "srcpkgs"                           
#> [203] "srcvers"                            "startsWith"                        
#> [205] "status"                             "stderr"                            
#> [207] "stop"                               "strsplit"                          
#> [209] "strwrap"                            "sub"                               
#> [211] "sum"                                "suppressWarnings"                  
#> [213] "Sys.getenv"                         "Sys.getpid"                        
#> [215] "Sys.glob"                           "Sys.setenv"                        
#> [217] "Sys.which"                          "system"                            
#> [219] "system2"                            "tcltk"                             
#> [221] "tempdir"                            "tempfile"                          
#> [223] "this"                               "timeout"                           
#> [225] "tlim"                               "tlim_cmd"                          
#> [227] "tmpd"                               "tmpd2"                             
#> [229] "tools"                              "try"                               
#> [231] "tryCatch"                           "tss"                               
#> [233] "type"                               "type2"                             
#> [235] "unique"                             "unlink"                            
#> [237] "unlist"                             "update"                            
#> [239] "upkgs"                              "urls"                              
#> [241] "userdir"                            "valid_package_version"             
#> [243] "verbose"                            "warning"                           
#> [245] "web"                                "writeLines"                        

# Does not find names of arguments
find_symbols(quote(foo(a = A)))
#> [1] "A"   "foo"
#> [1] "A"   "foo"

find_symbols(quote(function(x = X) NULL))
#> [1] "function" "X"       
#> [1] "function" "X"