Find all symbols used in a function or language object.

Find all symbols used in a function or language object.

find_symbols(x)

Arguments

x

A language object or function.

Value

A character vector of the symbols used in the object.

Examples

# Find all symbols used in the lm function
find_symbols(utils::install.packages)
#>   [1] "!"                                  "!="                                
#>   [3] "$"                                  "%in%"                              
#>   [5] "&"                                  "&&"                                
#>   [7] "("                                  "..."                               
#>   [9] ".Library"                           ".Platform"                         
#>  [11] ".R_LIBS"                            ".TkUp"                             
#>  [13] ".find_install_order"                ".install.macbinary"                
#>  [15] ".install.winbinary"                 ".libPaths"                         
#>  [17] ".make_dependency_list"              ".standard_regexps"                 
#>  [19] "::"                                 ":::"                               
#>  [21] "<"                                  "<-"                                
#>  [23] "=="                                 ">"                                 
#>  [25] "DL"                                 "GUI"                               
#>  [27] "INSTALL_opts"                       "NROW"                              
#>  [29] "Ncpus"                              "OK"                                
#>  [31] "OS.type"                            "R.home"                            
#>  [33] "R.version"                          "Sys.getenv"                        
#>  [35] "Sys.getpid"                         "Sys.glob"                          
#>  [37] "Sys.setenv"                         "Sys.which"                         
#>  [39] "["                                  "[["                                
#>  [41] "aDL"                                "action"                            
#>  [43] "all"                                "ans"                               
#>  [45] "any"                                "args"                              
#>  [47] "args0"                              "as.numeric_version"                
#>  [49] "askYesNo"                           "av"                                
#>  [51] "av1"                                "av2"                               
#>  [53] "available"                          "available.packages"                
#>  [55] "available_packages_filters_db"      "available_packages_filters_default"
#>  [57] "basename"                           "bins"                              
#>  [59] "binvers"                            "c"                                 
#>  [61] "capabilities"                       "cat"                               
#>  [63] "cbind"                              "character"                         
#>  [65] "clean"                              "close"                             
#>  [67] "cmd"                                "cmd0"                              
#>  [69] "colnames"                           "config"                            
#>  [71] "configure.args"                     "configure.vars"                    
#>  [73] "conn"                               "contrib.url"                       
#>  [75] "contriburl"                         "contriburl2"                       
#>  [77] "cwd"                                "data.frame"                        
#>  [79] "dependencies"                       "deps"                              
#>  [81] "destdir"                            "destfile"                          
#>  [83] "df"                                 "dir"                               
#>  [85] "dir.create"                         "dir.exists"                        
#>  [87] "dirname"                            "dirs"                              
#>  [89] "download.file"                      "download.packages"                 
#>  [91] "e"                                  "endsWith"                          
#>  [93] "env"                                "f"                                 
#>  [95] "f2"                                 "failed"                            
#>  [97] "fil"                                "file"                              
#>  [99] "file.access"                        "file.copy"                         
#> [101] "file.exists"                        "file.path"                         
#> [103] "files"                              "filters"                           
#> [105] "flush.console"                      "fn"                                
#> [107] "for"                                "found"                             
#> [109] "foundpkgs"                          "function"                          
#> [111] "getConfigureArgs"                   "getConfigureVars"                  
#> [113] "getDependencies"                    "getDeps"                           
#> [115] "getOption"                          "get_install_opts"                  
#> [117] "get_package_name"                   "get_timeout"                       
#> [119] "gettext"                            "gettextf"                          
#> [121] "getwd"                              "grep"                              
#> [123] "grepl"                              "gsub"                              
#> [125] "hasArchs"                           "hasSrc"                            
#> [127] "have_spaces"                        "i"                                 
#> [129] "if"                                 "inherits"                          
#> [131] "interactive"                        "inuse"                             
#> [133] "invisible"                          "is.character"                      
#> [135] "is.list"                            "is.logical"                        
#> [137] "is.na"                              "is.null"                           
#> [139] "isTRUE"                             "keep_outputs"                      
#> [141] "later"                              "length"                            
#> [143] "lib"                                "libpath"                           
#> [145] "libs_only"                          "libs_used"                         
#> [147] "list"                               "make.packages.html"                
#> [149] "match"                              "message"                           
#> [151] "method"                             "mfile"                             
#> [153] "missing"                            "msg"                               
#> [155] "na"                                 "names"                             
#> [157] "needsCmp"                           "next"                              
#> [159] "ngettext"                           "nonlocalrepos"                     
#> [161] "normalizePath"                      "nrow"                              
#> [163] "nzchar"                             "ok"                                
#> [165] "oldrlibs"                           "on.exit"                           
#> [167] "out"                                "outdir"                            
#> [169] "outfile"                            "outfiles"                          
#> [171] "output"                             "p"                                 
#> [173] "p0"                                 "paste"                             
#> [175] "paste0"                             "path.expand"                       
#> [177] "path.sep"                           "pkg"                               
#> [179] "pkgType"                            "pkgnames"                          
#> [181] "pkgs"                               "platform"                          
#> [183] "print"                              "quiet"                             
#> [185] "readLines"                          "repos"                             
#> [187] "res"                                "return"                            
#> [189] "row.names"                          "rownames"                          
#> [191] "s2"                                 "sQuote"                            
#> [193] "search"                             "select.list"                       
#> [195] "seq_len"                            "setdiff"                           
#> [197] "setwd"                              "shQuote"                           
#> [199] "shortPathName"                      "sort"                              
#> [201] "sort.list"                          "sprintf"                           
#> [203] "srcOnly"                            "srcpkgs"                           
#> [205] "srcvers"                            "startsWith"                        
#> [207] "status"                             "stderr"                            
#> [209] "stop"                               "strsplit"                          
#> [211] "strwrap"                            "sub"                               
#> [213] "sum"                                "suppressWarnings"                  
#> [215] "system"                             "system2"                           
#> [217] "tcltk"                              "tempdir"                           
#> [219] "tempfile"                           "this"                              
#> [221] "timeout"                            "tlim"                              
#> [223] "tlim_cmd"                           "tmpd"                              
#> [225] "tmpd2"                              "tools"                             
#> [227] "try"                                "tryCatch"                          
#> [229] "tss"                                "type"                              
#> [231] "type2"                              "unique"                            
#> [233] "unlink"                             "unlist"                            
#> [235] "update"                             "upkgs"                             
#> [237] "urls"                               "userdir"                           
#> [239] "valid_package_version"              "verbose"                           
#> [241] "warning"                            "web"                               
#> [243] "writeLines"                         "{"                                 
#> [245] "|"                                  "||"                                

# Does not find names of arguments
find_symbols(quote(foo(a = A)))
#> [1] "A"   "foo"
#> [1] "A"   "foo"

find_symbols(quote(function(x = X) NULL))
#> [1] "X"        "function"
#> [1] "function" "X"